case 5 energy solution/requirements.txt
# Python 3.11 required
# Energy Management Solution - Dependencies

# Core data science
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Anomaly detection
pyod>=1.1.0
scikit-learn>=1.3.0

# SHAP explainability
shap>=0.42.0

# Forecasting
timesfm[torch] @ git+https://github.com/google-research/timesfm.git
neuralforecast>=1.6.0
torch>=2.0.0

# LLM API
openai>=1.0.0

# Utilities
joblib>=1.3.0
python-dotenv>=1.0.0
